fixup save changes iwent

+++ app/src/main/java/com/example/gtamapirl/data/EventData.kt
@@ -9,14 +9,14 @@ class EventData {
     var description: String? = null
     var latitude: Double? = null
     var longitude: Double? = null
-    var participants: Int? = null
+    var participants: Long? = null
     var iconName: String? = null
 
     constructor()
 
     constructor(id: String, name: String, creator: String, date: String, time: String,
                 latitude: Double, longitude: Double, description: String,
-                participants: Int, iconName: String) {
+                participants: Long, iconName: String) {
         this.id = id
         this.name = name
         this.creator = creator
